Resolving css issues with previous commit
Previous commit(s) did not work with css of the website, therefore this commit reverses that work. Also moved movies.png back from images/deprecated/. This will NOT work without a compatible database, once we get a common schema and it is imported to our local db, we'll add the additional queries (for songs, people, etc) into movie_info.php.